moved isotope_storage.py into app/ dir
